pyproject.toml
[tool.poetry]
name = "party-parrot"
version = "0.1.0"
description = ""
authors = ["David Mack"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
PyAudio = "*"
DMXEnttecPro = "^0.4"
matplotlib = "^3.9.1"
numpy = "^2.0.1"
scipy = "^1.2.0"
events = "^0.5"
tk = "^0.1.0"
black = "^24.4.2"
pyenttec = "^1.4"
flask = "^3.1.0"
colorama = "^0.4.6"
pytest = "^8.3.5"
pillow = "^10.0.0"
moderngl = "^5.8.2"
fonttools = "^4.47.0"
pygame = "^2.5.0"
av = "11.0.0"
skia-python = "^138.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false
